Extract interface for TurnCostExtension, add no-op implementation
